change opmd into component edit stdio, freertos posix api, and px4 common

+++ boards/fmu-v2/board_config.h
@@ -13,9 +13,9 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-#define APP_LOAD_ADDRESS     (0x08004000)
-#define HSE_VALUE             ((uint32_t)24000000)
-#define LSE_VALUE             ((uint32_t)32768)
+#define APP_LOAD_ADDRESS      (0x08004000)
+#define HSE_VALUE             (24000000UL)
+#define LSE_VALUE             (32768UL)
 #define __FPU_PRESENT         1
 #define __FPU_USED            1
